requirements.txt
pyRAPL
cobs
pyserial
protobuf
psutil
cputemp
